WMV-366 & WMV-364: Continued work on resolving crash when loading model. (Not yet resolved) -Fixed a M2 reading bug involving Vec3D and double.

+++ src/games/wow/HardDriveFile.cpp
@@ -44,7 +44,7 @@ bool HardDriveFile::isAlreadyOpened()
     return false;
 }
 
-bool HardDriveFile::getFileSize(size_t & s)
+bool HardDriveFile::getFileSize(unsigned int & s)
 {
   if (!file && !file->isOpen())
     return false;
