+++ src/games/wow/WMOGroup.cpp
@@ -351,7 +351,7 @@ void WMOGroup::initDisplayList()
       size_t a = indices[batch->indexStart + i];
       for (size_t j = batch->vertexStart; j <= batch->vertexEnd; j++){
         if (vertices[a] == vertices[j]){
-          IndiceToVerts[batch->indexStart + i] = j;
+          IndiceToVerts[batch->indexStart + (uint)i] = (uint)j;
           break;
         }
       }
@@ -431,7 +431,7 @@ void WMOGroup::initLighting(int nLR, short *useLights)
   // "real" lighting?
   if ((flags & 0x2000) && hascv) {
 
-    Vec3D dirmin(1, 1, 1);
+    Vec3D dirmin(1.0, 1.0, 1.0);
     float lenmin;
     int lmin;
